pyproject.toml
[project]
name = "brew-history"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["pyyaml"]

[project.optional-dependencies]
dev = ["mypy", "types-pyyaml"]

[tool.mypy]
strict = true
files = ["scripts"]

[tool.ruff]
include = ["scripts/**/*.py"]
